Fixed a number of compiler warnings when compiling for Win32.
git-svn-id: http://www.bci2000.org/svn/trunk@3787 7ef59cdf-0624-0410-a474-e34ce222aafa

+++ src/contrib/SignalSource/RDAClient/RDAClientADC.cpp
@@ -93,7 +93,7 @@ RDAClientADC::Preflight( const SignalProperties&,
              << " (" << numInputChannels << ") "
              << endl;
     else
-      for( int i = 0; i < numInputChannels - 1; ++i )
+      for( size_t i = 0; i < numInputChannels - 1; ++i )
         goodOffsets &= ( Parameter( "SourceChOffset" )( i ) == 0 );
 
     if( !goodOffsets )
@@ -109,7 +109,7 @@ RDAClientADC::Preflight( const SignalProperties&,
              << " (" << numInputChannels << ") "
              << endl;
     else
-      for( int i = 0; i < numInputChannels - 1; ++i )
+      for( size_t i = 0; i < numInputChannels - 1; ++i )
       {
         double gain = preflightQueue.info().channelResolutions[ i ];
         double prmgain = Parameter( "SourceChGain")( i );
